wifi_rx: add webserver functionality to know more about the states of the buttons and add option to ignore the button release
For the supervisor of the escape room it could be highly important to know what buttons the people pressed and in which sequence. To let him have access to this information a webserver is added. It also allows the button sequence to be overridden. By accessing the right webpage (namely: 192.168.0.1/activate) the pin could be set to 1 from remote.

Additionally you might not want to take into account the button releasing but instead only the pressing and do not care too much about when the buttons are released. This is possible now with defining #IGNORE_BTN_RELEASE.